add: logic to create brand
